Factor out iter_files into its own module covered_files
