+++ src/reuse/project.py
@@ -34,13 +34,7 @@ from typing import (
 
 from binaryornot.check import is_binary
 
-from . import (
-    _IGNORE_DIR_PATTERNS,
-    _IGNORE_FILE_PATTERNS,
-    _IGNORE_MESON_PARENT_DIR_PATTERNS,
-    IdentifierNotFound,
-    ReuseInfo,
-)
+from . import IdentifierNotFound, ReuseInfo
 from ._licenses import EXCEPTION_MAP, LICENSE_MAP
 from ._util import (
     _LICENSEREF_PATTERN,
@@ -50,6 +44,7 @@ from ._util import (
     relative_from_root,
     reuse_info_of_file,
 )
+from .covered_files import iter_files
 from .global_licensing import (
     GlobalLicensing,
     NestedReuseTOML,
@@ -473,112 +468,3 @@ class Project:
             ).format(root)
         )
         return VCSStrategyNone(root)
-
-
-def is_path_ignored(
-    path: Path,
-    subset_files: Optional[Collection[StrPath]] = None,
-    include_submodules: bool = False,
-    include_meson_subprojects: bool = False,
-    vcs_strategy: Optional[VCSStrategy] = None,
-) -> bool:
-    """Is *path* ignored by some mechanism?"""
-    # pylint: disable=too-many-return-statements,too-many-branches
-    name = path.name
-    parent_parts = path.parent.parts
-    parent_dir = parent_parts[-1] if len(parent_parts) > 0 else ""
-
-    if path.is_symlink():
-        _LOGGER.debug("skipping symlink '%s'", path)
-        return True
-
-    if path.is_file():
-        if subset_files is not None and path.resolve() not in subset_files:
-            continue
-        for pattern in _IGNORE_FILE_PATTERNS:
-            if pattern.match(name):
-                return True
-        # Suppressing this error because I simply don't want to deal
-        # with that here.
-        with contextlib.suppress(OSError):
-            if path.stat().st_size == 0:
-                _LOGGER.debug("skipping 0-sized file '%s'", path)
-                return True
-
-    elif path.is_dir():
-        if subset_files is not None and not any(
-            is_relative_to(file_, path.resolve()) for file_ in subset_files
-        ):
-            return True
-        for pattern in _IGNORE_DIR_PATTERNS:
-            if pattern.match(name):
-                return True
-        if not include_meson_subprojects:
-            for pattern in _IGNORE_MESON_PARENT_DIR_PATTERNS:
-                if pattern.match(parent_dir):
-                    _LOGGER.info(
-                        "ignoring '%s' because it is a Meson subproject", path
-                    )
-                    return True
-        if (
-            not include_submodules
-            and vcs_strategy
-            and vcs_strategy.is_submodule(path)
-        ):
-            _LOGGER.info("ignoring '%s' because it is a submodule", path)
-            return True
-
-    if vcs_strategy and vcs_strategy.is_ignored(path):
-        return True
-
-    return False
-
-
-def iter_files(
-    directory: StrPath,
-    subset_files: Optional[Collection[StrPath]] = None,
-    include_submodules: bool = False,
-    include_meson_subprojects: bool = False,
-    vcs_strategy: Optional[VCSStrategy] = None,
-) -> Generator[Path, None, None]:
-    """Yield all Covered Files in *directory* and its subdirectories according
-    to the REUSE Specification.
-    """
-    directory = Path(directory)
-    if subset_files is not None:
-        subset_files = cast(
-            Set[Path], {Path(file_).resolve() for file_ in subset_files}
-        )
-
-    for root_str, dirs, files in os.walk(directory):
-        root = Path(root_str)
-        _LOGGER.debug("currently walking in '%s'", root)
-
-        # Don't walk ignored directories
-        for dir_ in list(dirs):
-            the_dir = root / dir_
-            if is_path_ignored(
-                the_dir,
-                subset_files=subset_files,
-                include_submodules=include_submodules,
-                include_meson_subprojects=include_meson_subprojects,
-                vcs_strategy=vcs_strategy,
-            ):
-                _LOGGER.debug("ignoring '%s'", the_dir)
-                dirs.remove(dir_)
-
-        # Filter files.
-        for file_ in files:
-            the_file = root / file_
-            if is_path_ignored(
-                the_file,
-                subset_files=subset_files,
-                include_submodules=include_submodules,
-                include_meson_subprojects=include_meson_subprojects,
-                vcs_strategy=vcs_strategy,
-            ):
-                _LOGGER.debug("ignoring '%s'", the_file)
-                continue
-
-            _LOGGER.debug("yielding '%s'", the_file)
-            yield the_file
